🎨 Modularized and 📝 documented search.js
